+++ src/problem_builders/frequency_domain_problem_builder.cpp
@@ -1,0 +1,10 @@
+#include "frequency_domain_problem_builder.hpp"
+
+namespace hephaestus {
+FrequencyDomainProblem::FrequencyDomainProblem(
+    const hephaestus::InputParameters &params)
+    : Problem(params),
+      formulation(
+          params.GetParam<hephaestus::SteadyFormulation *>("Formulation")) {}
+
+} // namespace hephaestus
