+++ src/problem_builders/time_domain_problem_builder.cpp
@@ -1,0 +1,9 @@
+#include "time_domain_problem_builder.hpp"
+
+namespace hephaestus {
+TransientProblem::TransientProblem(const hephaestus::InputParameters &params)
+    : Problem(params),
+      formulation(
+          params.GetParam<hephaestus::TransientFormulation *>("Formulation")) {}
+
+} // namespace hephaestus
